pyproject.toml
[tool.poetry]
name = "furniflow-ai-api"
version = "1.0.0"
description = "Artificial Inteligence API for Classificate Product Types on Furniflow"
authors = ["Matheus Oliveira dos Santos", "math.olsantos@gmail.com"]
readme = "README.md"
repository = "https://github.com/mathsantos44/furnflow-ai-api"

[information]
contact = ["Matheus Oliveira dos Santos", "math.olsantos@gmail.com"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
joblib = "^1.3.2"
pandas = "^2.2.1"
numpy = "^1.26.4"
scikit-learn = "^1.4.1.post1"
